+++ php/classes/JSONDatabase.php
@@ -476,7 +476,7 @@ class JSONDatabase {
     }
 
     // MANDATORY REFERENCE to edit directly: PHP 5+
-    private function _edit(&$obj, $editObj): bool {
+    private function _edit(&$content, $editObj): bool {
         // must be associative array
         $editObjType = gettype($editObj);
         if (is_primitive($editObj) || array_sequential($editObj))
@@ -493,7 +493,7 @@ class JSONDatabase {
             throw new HTTPException('ID must be a string or number', 400);
 
         // object not found
-        if (!array_key_exists($id, $obj->content) || !check($obj->content[$id]))
+        if (!array_key_exists($id, $content) || !check($content[$id]))
             throw new HTTPException('ID doesn\'t exist in collection', 400);
 
         // field required
@@ -527,36 +527,36 @@ class JSONDatabase {
         // field not needed for set or push operation (can create fields)
         // missing field in remove doesn't matter since it's gone either way
         if (
-            !isset($obj->content[$id][$field]) and
+            !isset($content[$id][$field]) and
             ($operation != 'set' and $operation != 'remove' and $operation != 'array-push')
         )
             throw new HTTPException("Field $field doesn't exist in ID $id", 400);
 
         switch ($operation) {
             case 'set':
-                $obj->content[$id][$field] = $value;
+                $content[$id][$field] = $value;
                 return true;
             case 'remove':
-                unset($obj->content[$id][$field]);
+                unset($content[$id][$field]);
                 return true;
             case 'append':
                 // check type string
-                if (gettype($obj->content[$id][$field]) != 'string' or gettype($value) != 'string')
+                if (gettype($content[$id][$field]) != 'string' or gettype($value) != 'string')
                     throw new HTTPException('append requires string values', 400);
 
-                $obj->content[$id][$field] .= $value;
+                $content[$id][$field] .= $value;
                 return true;
             case 'invert':
                 // check type boolean
-                if (gettype($obj->content[$id][$field]) != 'boolean')
+                if (gettype($content[$id][$field]) != 'boolean')
                     throw new HTTPException('invert field must be a boolean', 400);
 
-                $obj->content[$id][$field] = !$obj->content[$id][$field];
+                $content[$id][$field] = !$content[$id][$field];
                 return true;
             case 'increment':
             case 'decrement':
                 // check type number
-                if (!is_number_like($obj->content[$id][$field]))
+                if (!is_number_like($content[$id][$field]))
                     throw new HTTPException('increment and decrement fields must be numbers', 400);
 
                 $change = $operation == 'increment' ? 1 : -1;
@@ -571,29 +571,29 @@ class JSONDatabase {
                         throw new HTTPException('increment and decrement values must be numbers', 400);
                 }
 
-                $obj->content[$id][$field] += $change;
+                $content[$id][$field] += $change;
                 return true;
             case 'array-push':
                 // create it if not here
-                if (!isset($obj->content[$id][$field]))
-                    $obj->content[$id][$field] = [];
+                if (!isset($content[$id][$field]))
+                    $content[$id][$field] = [];
 
                 // check if our field array
                 if (
-                    gettype($obj->content[$id][$field]) != 'array' ||
-                    array_assoc($obj->content[$id][$field])
+                    gettype($content[$id][$field]) != 'array' ||
+                    array_assoc($content[$id][$field])
                 )
                     throw new HTTPException('array-push field must be an array', 400);
 
-                array_push($obj->content[$id][$field], $value);
+                array_push($content[$id][$field], $value);
 
                 return true;
 
             case 'array-delete':
                 // check if our field array
                 if (
-                    gettype($obj->content[$id][$field]) != 'array' ||
-                    array_assoc($obj->content[$id][$field])
+                    gettype($content[$id][$field]) != 'array' ||
+                    array_assoc($content[$id][$field])
                 )
                     throw new HTTPException('array-delete field must be an array', 400);
 
@@ -601,11 +601,11 @@ class JSONDatabase {
                 if (gettype($value) != 'integer')
                     throw new HTTPException('array-delete value must be a number', 400);
 
-                array_splice($obj->content[$id][$field], $value, 1);
+                array_splice($content[$id][$field], $value, 1);
 
                 return true;
             case 'array-splice':
-                if (array_assoc($obj->content[$id][$field]))
+                if (array_assoc($content[$id][$field]))
                     throw new HTTPException('array-splice field must be an array', 400);
 
                 // value must be an array starting with two integers
@@ -618,9 +618,9 @@ class JSONDatabase {
                     throw new HTTPException('Incorrect array-splice options', 400);
 
                 if (count($value) > 2)
-                    array_splice($obj->content[$id][$field], $value[0], $value[1], $value[2]);
+                    array_splice($content[$id][$field], $value[0], $value[1], $value[2]);
                 else
-                    array_splice($obj->content[$id][$field], $value[0], $value[1]);
+                    array_splice($content[$id][$field], $value[0], $value[1]);
 
                 return true;
             default:
@@ -632,7 +632,7 @@ class JSONDatabase {
 
     public function editField($editObj) {
         $fileObj = $this->read(true);
-        $this->_edit($fileObj, $editObj);
+        $this->_edit($fileObj->content, $editObj);
         $this->write($fileObj);
     }